feat: move cml related types to cml package
