+++ packages/cml/src/index.ts
@@ -1,1 +1,2 @@
 export * from './getRgxInstance';
+export * from './types';
