+++ packages/nextjs-contentful-renderer/src/utils/contentful/getPageById/types.ts
@@ -1,4 +1,5 @@
 import { Theme } from '@cmpsr/components';
+import { Block } from '@cmpsr/cml';
 
 export interface Page {
   id: string;
@@ -18,26 +19,4 @@ interface PageMeta {
   propertyName: string;
   propertyValue: string;
   content: string;
-}
-
-export interface Block {
-  models: Model[];
-  propsValues: PropsValue[];
-}
-
-export type Model = ResponsiveValue<string>;
-
-export type PropsValue = ResponsiveValue<
-  Record<
-    string,
-    string | number | boolean | Record<string, string | string[] | number | boolean | number[] | boolean[]>
-  >
->;
-
-export interface ResponsiveValue<T> {
-  base?: T;
-  md?: T;
-  lg?: T;
-  xl?: T;
-  xxl?: T;
 }
