+++ packages/nextjs-contentful-renderer/src/utils/getPageById/configFooter/configFooter.ts
@@ -1,4 +1,5 @@
+import { Block } from '@cmpsr/cml';
 import { insertBlock } from '../insertBlock';
-import { Block, Page } from '../../contentful/getPageById/types';
+import { Page } from '../../contentful/getPageById/types';
 
 export const configFooter = (page: Partial<Page>): Block[] => insertBlock(page.content, page.footer?.[0], 'push');
