+++ packages/nextjs-contentful-renderer/src/utils/getPageById/configNavbar/configNavbar.ts
@@ -1,4 +1,5 @@
-import { Block, Page } from '../../contentful/getPageById/types';
+import { Block } from '@cmpsr/cml';
+import { Page } from '../../contentful/getPageById/types';
 import { insertBlock } from '../insertBlock';
 
 export const configNavbar = (page: Partial<Page>): Block[] => insertBlock(page.content, page.navbar?.[0], 'unshift');
