+++ packages/nextjs-contentful-renderer/src/utils/getPageById/insertBlock/insertBlock.ts
@@ -1,4 +1,4 @@
-import { Block } from '../../contentful/getPageById/types';
+import { Block } from '@cmpsr/cml';
 
 export const insertBlock = (originalContent: Block[], block: Block, method: 'unshift' | 'push') => {
   if (!block) return originalContent;
